+++ docs/src/components/Docs/Props/CustomizingToolbarProp/defaultToolbar.js
@@ -1,136 +1,137 @@
-export default '{\n' +
-  '  options: [\'inline\', \'blockType\', \'fontSize\', \'fontFamily\', \'list\', \'textAlign\', \'colorPicker\', \'link\', \'embedded\', \'emoji\', \'image\', \'remove\', \'history\'],\n' +
-  '  inline: {\n' +
-  '    inDropdown: false,\n' +
-  '    className: undefined,\n' +
-  '    component: undefined,\n' +
-  '    dropdownClassName: undefined,\n' +
-  '    options: [\'bold\', \'italic\', \'underline\', \'strikethrough\', \'monospace\', \'superscript\', \'subscript\'],\n' +
-  '    bold: { icon: bold, className: undefined },\n' +
-  '    italic: { icon: italic, className: undefined },\n' +
-  '    underline: { icon: underline, className: undefined },\n' +
-  '    strikethrough: { icon: strikethrough, className: undefined },\n' +
-  '    monospace: { icon: monospace, className: undefined },\n' +
-  '    superscript: { icon: superscript, className: undefined },\n' +
-  '    subscript: { icon: subscript, className: undefined },\n' +
-  '  },\n' +
-  '  blockType: {\n' +
-  '    inDropdown: true,\n' +
-  '    options: [\'Normal\', \'H1\', \'H2\', \'H3\', \'H4\', \'H5\', \'H6\', \'Blockquote\', \'Code\'],\n' +
-  '    className: undefined,\n' +
-  '    component: undefined,\n' +
-  '    dropdownClassName: undefined,\n' +
-  '  },\n' +
-  '  fontSize: {\n' +
-  '    icon: fontSize,\n' +
-  '    options: [8, 9, 10, 11, 12, 14, 16, 18, 24, 30, 36, 48, 60, 72, 96],\n' +
-  '    className: undefined,\n' +
-  '    component: undefined,\n' +
-  '    dropdownClassName: undefined,\n' +
-  '  },\n' +
-  '  fontFamily: {\n' +
-  '    options: [\'Arial\', \'Georgia\', \'Impact\', \'Tahoma\', \'Times New Roman\', \'Verdana\'],\n' +
-  '    className: undefined,\n' +
-  '    component: undefined,\n' +
-  '    dropdownClassName: undefined,\n' +
-  '  },\n' +
-  '  list: {\n' +
-  '    inDropdown: false,\n' +
-  '    className: undefined,\n' +
-  '    component: undefined,\n' +
-  '    dropdownClassName: undefined,\n' +
-  '    options: [\'unordered\', \'ordered\', \'indent\', \'outdent\'],\n' +
-  '    unordered: { icon: unordered, className: undefined },\n' +
-  '    ordered: { icon: ordered, className: undefined },\n' +
-  '    indent: { icon: indent, className: undefined },\n' +
-  '    outdent: { icon: outdent, className: undefined },\n' +
-  '  },\n' +
-  '  textAlign: {\n' +
-  '    inDropdown: false,\n' +
-  '    className: undefined,\n' +
-  '    component: undefined,\n' +
-  '    dropdownClassName: undefined,\n' +
-  '    options: [\'left\', \'center\', \'right\', \'justify\'],\n' +
-  '    left: { icon: left, className: undefined },\n' +
-  '    center: { icon: center, className: undefined },\n' +
-  '    right: { icon: right, className: undefined },\n' +
-  '    justify: { icon: justify, className: undefined },\n' +
-  '  },\n' +
-  '  colorPicker: {\n' +
-  '    icon: color,\n' +
-  '    className: undefined,\n' +
-  '    component: undefined,\n' +
-  '    popupClassName: undefined,\n' +
-  '    colors: [\'rgb(97,189,109)\', \'rgb(26,188,156)\', \'rgb(84,172,210)\', \'rgb(44,130,201)\',\n' +
-  '      \'rgb(147,101,184)\', \'rgb(71,85,119)\', \'rgb(204,204,204)\', \'rgb(65,168,95)\', \'rgb(0,168,133)\',\n' +
-  '      \'rgb(61,142,185)\', \'rgb(41,105,176)\', \'rgb(85,57,130)\', \'rgb(40,50,78)\', \'rgb(0,0,0)\',\n' +
-  '      \'rgb(247,218,100)\', \'rgb(251,160,38)\', \'rgb(235,107,86)\', \'rgb(226,80,65)\', \'rgb(163,143,132)\',\n' +
-  '      \'rgb(239,239,239)\', \'rgb(255,255,255)\', \'rgb(250,197,28)\', \'rgb(243,121,52)\', \'rgb(209,72,65)\',\n' +
-  '      \'rgb(184,49,47)\', \'rgb(124,112,107)\', \'rgb(209,213,216)\'],\n' +
-  '  },\n' +
-  '  link: {\n' +
-  '    inDropdown: false,\n' +
-  '    className: undefined,\n' +
-  '    component: undefined,\n' +
-  '    popupClassName: undefined,\n' +
-  '    dropdownClassName: undefined,\n' +
-  '    showOpenOptionOnHover: true,\n' +
-  '    defaultTargetOption: \'_self\',\n' +
-  '    options: [\'link\', \'unlink\'],\n' +
-  '    link: { icon: link, className: undefined },\n' +
-  '    unlink: { icon: unlink, className: undefined },\n' +
-  '  },\n' +
-  '  emoji: {\n' +
-  '    icon: emoji,\n' +
-  '    className: undefined,\n' +
-  '    component: undefined,\n' +
-  '    popupClassName: undefined,\n' +
-  '    emojis: [\n' +
-  '      \'😀\', \'😁\', \'😂\', \'😃\', \'😉\', \'😋\', \'😎\', \'😍\', \'😗\', \'🤗\', \'🤔\', \'😣\', \'😫\', \'😴\', \'😌\', \'🤓\',\n' +
-  '      \'😛\', \'😜\', \'😠\', \'😇\', \'😷\', \'😈\', \'👻\', \'😺\', \'😸\', \'😹\', \'😻\', \'😼\', \'😽\', \'🙀\', \'🙈\',\n' +
-  '      \'🙉\', \'🙊\', \'👼\', \'👮\', \'🕵\', \'💂\', \'👳\', \'🎅\', \'👸\', \'👰\', \'👲\', \'🙍\', \'🙇\', \'🚶\', \'🏃\', \'💃\',\n' +
-  '      \'⛷\', \'🏂\', \'🏌\', \'🏄\', \'🚣\', \'🏊\', \'⛹\', \'🏋\', \'🚴\', \'👫\', \'💪\', \'👈\', \'👉\', \'👉\', \'👆\', \'🖕\',\n' +
-  '      \'👇\', \'🖖\', \'🤘\', \'🖐\', \'👌\', \'👍\', \'👎\', \'✊\', \'👊\', \'👏\', \'🙌\', \'🙏\', \'🐵\', \'🐶\', \'🐇\', \'🐥\',\n' +
-  '      \'🐸\', \'🐌\', \'🐛\', \'🐜\', \'🐝\', \'🍉\', \'🍄\', \'🍔\', \'🍤\', \'🍨\', \'🍪\', \'🎂\', \'🍰\', \'🍾\', \'🍷\', \'🍸\',\n' +
-  '      \'🍺\', \'🌍\', \'🚑\', \'⏰\', \'🌙\', \'🌝\', \'🌞\', \'⭐\', \'🌟\', \'🌠\', \'🌨\', \'🌩\', \'⛄\', \'🔥\', \'🎄\', \'🎈\',\n' +
-  '      \'🎉\', \'🎊\', \'🎁\', \'🎗\', \'🏀\', \'🏈\', \'🎲\', \'🔇\', \'🔈\', \'📣\', \'🔔\', \'🎵\', \'🎷\', \'💰\', \'🖊\', \'📅\',\n' +
-  '      \'✅\', \'❎\', \'💯\',\n' +
-  '    ],\n' +
-  '  },\n' +
-  '  embedded: {\n' +
-  '    icon: embedded,\n' +
-  '    className: undefined,\n' +
-  '    component: undefined,\n' +
-  '    popupClassName: undefined,\n' +
-  '    defaultSize: {\n' +
-  '      height: \'auto\',\n' +
-  '      width: \'auto\',\n' +
-  '    },\n' +
-  '  },\n' +
-  '  image: {\n' +
-  '    icon: image,\n' +
-  '    className: undefined,\n' +
-  '    component: undefined,\n' +
-  '    popupClassName: undefined,\n' +
-  '    urlEnabled: true,\n' +
-  '    uploadEnabled: true,\n' +
-  '    alignmentEnabled: true,\n' +
-  '    uploadCallback: undefined,\n' +
-  '    inputAccept: \'image/gif,image/jpeg,image/jpg,image/png,image/svg\',\n' +
-  '    alt: { present: false, mandatory: false },\n' +
-  '    defaultSize: {\n' +
-  '      height: \'auto\',\n' +
-  '      width: \'auto\',\n' +
-  '    },\n' +
-  '  },\n' +
-  '  remove: { icon: eraser, className: undefined, component: undefined },\n' +
-  '  history: {\n' +
-  '    inDropdown: false,\n' +
-  '    className: undefined,\n' +
-  '    component: undefined,\n' +
-  '    dropdownClassName: undefined,\n' +
-  '    options: [\'undo\', \'redo\'],\n' +
-  '    undo: { icon: undo, className: undefined },\n' +
-  '    redo: { icon: redo, className: undefined },\n' +
-  '  },\n' +
-  '}';
+export default "{\n" +
+  "  options: ['inline', 'blockType', 'fontSize', 'fontFamily', 'list', 'textAlign', 'colorPicker', 'link', 'embedded', 'emoji', 'image', 'remove', 'history'],\n" +
+  "  inline: {\n" +
+  "    inDropdown: false,\n" +
+  "    className: undefined,\n" +
+  "    component: undefined,\n" +
+  "    dropdownClassName: undefined,\n" +
+  "    options: ['bold', 'italic', 'underline', 'strikethrough', 'monospace', 'superscript', 'subscript'],\n" +
+  "    bold: { icon: bold, className: undefined },\n" +
+  "    italic: { icon: italic, className: undefined },\n" +
+  "    underline: { icon: underline, className: undefined },\n" +
+  "    strikethrough: { icon: strikethrough, className: undefined },\n" +
+  "    monospace: { icon: monospace, className: undefined },\n" +
+  "    superscript: { icon: superscript, className: undefined },\n" +
+  "    subscript: { icon: subscript, className: undefined },\n" +
+  "  },\n" +
+  "  blockType: {\n" +
+  "    inDropdown: true,\n" +
+  "    options: ['Normal', 'H1', 'H2', 'H3', 'H4', 'H5', 'H6', 'Blockquote', 'Code'],\n" +
+  "    className: undefined,\n" +
+  "    component: undefined,\n" +
+  "    dropdownClassName: undefined,\n" +
+  "  },\n" +
+  "  fontSize: {\n" +
+  "    icon: fontSize,\n" +
+  "    options: [8, 9, 10, 11, 12, 14, 16, 18, 24, 30, 36, 48, 60, 72, 96],\n" +
+  "    className: undefined,\n" +
+  "    component: undefined,\n" +
+  "    dropdownClassName: undefined,\n" +
+  "  },\n" +
+  "  fontFamily: {\n" +
+  "    options: ['Arial', 'Georgia', 'Impact', 'Tahoma', 'Times New Roman', 'Verdana'],\n" +
+  "    className: undefined,\n" +
+  "    component: undefined,\n" +
+  "    dropdownClassName: undefined,\n" +
+  "  },\n" +
+  "  list: {\n" +
+  "    inDropdown: false,\n" +
+  "    className: undefined,\n" +
+  "    component: undefined,\n" +
+  "    dropdownClassName: undefined,\n" +
+  "    options: ['unordered', 'ordered', 'indent', 'outdent'],\n" +
+  "    unordered: { icon: unordered, className: undefined },\n" +
+  "    ordered: { icon: ordered, className: undefined },\n" +
+  "    indent: { icon: indent, className: undefined },\n" +
+  "    outdent: { icon: outdent, className: undefined },\n" +
+  "  },\n" +
+  "  textAlign: {\n" +
+  "    inDropdown: false,\n" +
+  "    className: undefined,\n" +
+  "    component: undefined,\n" +
+  "    dropdownClassName: undefined,\n" +
+  "    options: ['left', 'center', 'right', 'justify'],\n" +
+  "    left: { icon: left, className: undefined },\n" +
+  "    center: { icon: center, className: undefined },\n" +
+  "    right: { icon: right, className: undefined },\n" +
+  "    justify: { icon: justify, className: undefined },\n" +
+  "  },\n" +
+  "  colorPicker: {\n" +
+  "    icon: color,\n" +
+  "    className: undefined,\n" +
+  "    component: undefined,\n" +
+  "    popupClassName: undefined,\n" +
+  "    colors: ['rgb(97,189,109)', 'rgb(26,188,156)', 'rgb(84,172,210)', 'rgb(44,130,201)',\n" +
+  "      'rgb(147,101,184)', 'rgb(71,85,119)', 'rgb(204,204,204)', 'rgb(65,168,95)', 'rgb(0,168,133)',\n" +
+  "      'rgb(61,142,185)', 'rgb(41,105,176)', 'rgb(85,57,130)', 'rgb(40,50,78)', 'rgb(0,0,0)',\n" +
+  "      'rgb(247,218,100)', 'rgb(251,160,38)', 'rgb(235,107,86)', 'rgb(226,80,65)', 'rgb(163,143,132)',\n" +
+  "      'rgb(239,239,239)', 'rgb(255,255,255)', 'rgb(250,197,28)', 'rgb(243,121,52)', 'rgb(209,72,65)',\n" +
+  "      'rgb(184,49,47)', 'rgb(124,112,107)', 'rgb(209,213,216)'],\n" +
+  "  },\n" +
+  "  link: {\n" +
+  "    inDropdown: false,\n" +
+  "    className: undefined,\n" +
+  "    component: undefined,\n" +
+  "    popupClassName: undefined,\n" +
+  "    dropdownClassName: undefined,\n" +
+  "    showOpenOptionOnHover: true,\n" +
+  "    defaultTargetOption: '_self',\n" +
+  "    options: ['link', 'unlink'],\n" +
+  "    link: { icon: link, className: undefined },\n" +
+  "    unlink: { icon: unlink, className: undefined },\n" +
+  "  },\n" +
+  "  emoji: {\n" +
+  "    icon: emoji,\n" +
+  "    className: undefined,\n" +
+  "    component: undefined,\n" +
+  "    popupClassName: undefined,\n" +
+  "    emojis: [\n" +
+  "      '😀', '😁', '😂', '😃', '😉', '😋', '😎', '😍', '😗', '🤗', '🤔', '😣', '😫', '😴', '😌', '🤓',\n" +
+  "      '😛', '😜', '😠', '😇', '😷', '😈', '👻', '😺', '😸', '😹', '😻', '😼', '😽', '🙀', '🙈',\n" +
+  "      '🙉', '🙊', '👼', '👮', '🕵', '💂', '👳', '🎅', '👸', '👰', '👲', '🙍', '🙇', '🚶', '🏃', '💃',\n" +
+  "      '⛷', '🏂', '🏌', '🏄', '🚣', '🏊', '⛹', '🏋', '🚴', '👫', '💪', '👈', '👉', '👉', '👆', '🖕',\n" +
+  "      '👇', '🖖', '🤘', '🖐', '👌', '👍', '👎', '✊', '👊', '👏', '🙌', '🙏', '🐵', '🐶', '🐇', '🐥',\n" +
+  "      '🐸', '🐌', '🐛', '🐜', '🐝', '🍉', '🍄', '🍔', '🍤', '🍨', '🍪', '🎂', '🍰', '🍾', '🍷', '🍸',\n" +
+  "      '🍺', '🌍', '🚑', '⏰', '🌙', '🌝', '🌞', '⭐', '🌟', '🌠', '🌨', '🌩', '⛄', '🔥', '🎄', '🎈',\n" +
+  "      '🎉', '🎊', '🎁', '🎗', '🏀', '🏈', '🎲', '🔇', '🔈', '📣', '🔔', '🎵', '🎷', '💰', '🖊', '📅',\n" +
+  "      '✅', '❎', '💯',\n" +
+  "    ],\n" +
+  "  },\n" +
+  "  embedded: {\n" +
+  "    icon: embedded,\n" +
+  "    className: undefined,\n" +
+  "    component: undefined,\n" +
+  "    popupClassName: undefined,\n" +
+  "    defaultSize: {\n" +
+  "      height: 'auto',\n" +
+  "      width: 'auto',\n" +
+  "    },\n" +
+  "  },\n" +
+  "  image: {\n" +
+  "    icon: image,\n" +
+  "    className: undefined,\n" +
+  "    component: undefined,\n" +
+  "    popupClassName: undefined,\n" +
+  "    urlEnabled: true,\n" +
+  "    uploadEnabled: true,\n" +
+  "    alignmentEnabled: true,\n" +
+  "    uploadCallback: undefined,\n" +
+  "    previewImage: false,\n" +
+  "    inputAccept: 'image/gif,image/jpeg,image/jpg,image/png,image/svg',\n" +
+  "    alt: { present: false, mandatory: false },\n" +
+  "    defaultSize: {\n" +
+  "      height: 'auto',\n" +
+  "      width: 'auto',\n" +
+  "    },\n" +
+  "  },\n" +
+  "  remove: { icon: eraser, className: undefined, component: undefined },\n" +
+  "  history: {\n" +
+  "    inDropdown: false,\n" +
+  "    className: undefined,\n" +
+  "    component: undefined,\n" +
+  "    dropdownClassName: undefined,\n" +
+  "    options: ['undo', 'redo'],\n" +
+  "    undo: { icon: undo, className: undefined },\n" +
+  "    redo: { icon: redo, className: undefined },\n" +
+  "  },\n" +
+  "}";