+++ src/Editor/index.js
@@ -1,41 +1,41 @@
 /* @flow */
 
-import React, { Component } from 'react';
-import PropTypes from 'prop-types';
+import React, { Component } from "react";
+import PropTypes from "prop-types";
 import {
   Editor,
   EditorState,
   RichUtils,
   convertToRaw,
   convertFromRaw,
-  CompositeDecorator,
-} from 'draft-js';
+  CompositeDecorator
+} from "draft-js";
 import {
   changeDepth,
   handleNewLine,
   blockRenderMap,
   getCustomStyleMap,
   extractInlineStyle,
-  getSelectedBlocksType,
-} from 'draftjs-utils';
-import classNames from 'classnames';
-import ModalHandler from '../event-handler/modals';
-import FocusHandler from '../event-handler/focus';
-import KeyDownHandler from '../event-handler/keyDown';
-import SuggestionHandler from '../event-handler/suggestions';
-import blockStyleFn from '../utils/BlockStyle';
-import { mergeRecursive } from '../utils/toolbar';
-import { hasProperty, filter } from '../utils/common';
-import { handlePastedText } from '../utils/handlePaste';
-import Controls from '../controls';
-import getLinkDecorator from '../decorators/Link';
-import getMentionDecorators from '../decorators/Mention';
-import getHashtagDecorator from '../decorators/HashTag';
-import getBlockRenderFunc from '../renderer';
-import defaultToolbar from '../config/defaultToolbar';
-import localeTranslations from '../i18n';
-import './styles.css';
-import '../../css/Draft.css';
+  getSelectedBlocksType
+} from "draftjs-utils";
+import classNames from "classnames";
+import ModalHandler from "../event-handler/modals";
+import FocusHandler from "../event-handler/focus";
+import KeyDownHandler from "../event-handler/keyDown";
+import SuggestionHandler from "../event-handler/suggestions";
+import blockStyleFn from "../utils/BlockStyle";
+import { mergeRecursive } from "../utils/toolbar";
+import { hasProperty, filter } from "../utils/common";
+import { handlePastedText } from "../utils/handlePaste";
+import Controls from "../controls";
+import getLinkDecorator from "../decorators/Link";
+import getMentionDecorators from "../decorators/Mention";
+import getHashtagDecorator from "../decorators/HashTag";
+import getBlockRenderFunc from "../renderer";
+import defaultToolbar from "../config/defaultToolbar";
+import localeTranslations from "../i18n";
+import "./styles.css";
+import "../../css/Draft.css";
 
 export default class WysiwygEditor extends Component {
   static propTypes = {
@@ -82,15 +82,16 @@ export default class WysiwygEditor extends Component {
     customBlockRenderFunc: PropTypes.func,
     wrapperId: PropTypes.number,
     customDecorators: PropTypes.array,
+    editorRef: PropTypes.func
   };
 
   static defaultProps = {
     toolbarOnFocus: false,
     toolbarHidden: false,
     stripPastedStyles: false,
-    localization: { locale: 'en', translations: {} },
-    customDecorators: [],
-  }
+    localization: { locale: "en", translations: {} },
+    customDecorators: []
+  };
 
   constructor(props) {
     super(props);
@@ -98,18 +99,23 @@ export default class WysiwygEditor extends Component {
     this.state = {
       editorState: undefined,
       editorFocused: false,
-      toolbar,
+      toolbar
     };
-    const wrapperId = props.wrapperId ? props.wrapperId : Math.floor(Math.random() * 10000);
+    const wrapperId = props.wrapperId
+      ? props.wrapperId
+      : Math.floor(Math.random() * 10000);
     this.wrapperId = `rdw-wrapper-${wrapperId}`;
     this.modalHandler = new ModalHandler();
     this.focusHandler = new FocusHandler();
-    this.blockRendererFn = getBlockRenderFunc({
-      isReadOnly: this.isReadOnly,
-      isImageAlignmentEnabled: this.isImageAlignmentEnabled,
-      getEditorState: this.getEditorState,
-      onChange: this.onChange,
-    }, props.customBlockRenderFunc);
+    this.blockRendererFn = getBlockRenderFunc(
+      {
+        isReadOnly: this.isReadOnly,
+        isImageAlignmentEnabled: this.isImageAlignmentEnabled,
+        getEditorState: this.getEditorState,
+        onChange: this.onChange
+      },
+      props.customBlockRenderFunc
+    );
     this.editorProps = this.filterEditorProps(props);
     this.customStyleMap = getCustomStyleMap();
   }
@@ -119,7 +125,7 @@ export default class WysiwygEditor extends Component {
     const editorState = this.createEditorState(this.compositeDecorator);
     extractInlineStyle(editorState);
     this.setState({
-      editorState,
+      editorState
     });
   }
 
@@ -134,16 +140,21 @@ export default class WysiwygEditor extends Component {
       const toolbar = mergeRecursive(defaultToolbar, props.toolbar);
       newState.toolbar = toolbar;
     }
-    if (hasProperty(props, 'editorState') && this.props.editorState !== props.editorState) {
+    if (
+      hasProperty(props, "editorState") &&
+      this.props.editorState !== props.editorState
+    ) {
       if (props.editorState) {
-        newState.editorState = EditorState.set(
-          props.editorState,
-          { decorator: this.compositeDecorator },
-        );
+        newState.editorState = EditorState.set(props.editorState, {
+          decorator: this.compositeDecorator
+        });
       } else {
         newState.editorState = EditorState.createEmpty(this.compositeDecorator);
       }
-    } else if (hasProperty(props, 'contentState') && this.props.contentState !== props.contentState) {
+    } else if (
+      hasProperty(props, "contentState") &&
+      this.props.contentState !== props.contentState
+    ) {
       if (props.contentState) {
         const newEditorState = this.changeEditorState(props.contentState);
         if (newEditorState) {
@@ -153,7 +164,10 @@ export default class WysiwygEditor extends Component {
         newState.editorState = EditorState.createEmpty(this.compositeDecorator);
       }
     }
-    if (props.editorState !== this.props.editorState || props.contentState !== this.props.contentState) {
+    if (
+      props.editorState !== this.props.editorState ||
+      props.contentState !== this.props.contentState
+    ) {
       extractInlineStyle(newState.editorState);
     }
     this.setState(newState);
@@ -163,28 +177,33 @@ export default class WysiwygEditor extends Component {
 
   onEditorBlur: Function = (): void => {
     this.setState({
-      editorFocused: false,
+      editorFocused: false
     });
   };
 
   onEditorFocus: Function = (event): void => {
     const { onFocus } = this.props;
     this.setState({
-      editorFocused: true,
+      editorFocused: true
     });
-    if (onFocus && this.focusHandler.isEditorFocused()) {
+    const editFocused = this.focusHandler.isEditorFocused();
+    if (onFocus && editFocused) {
       onFocus(event);
     }
   };
 
   onEditorMouseDown: Function = (): void => {
     this.focusHandler.onEditorMouseDown();
-  }
+  };
 
   onTab: Function = (event): boolean => {
     const { onTab } = this.props;
     if (!onTab || !onTab(event)) {
-      const editorState = changeDepth(this.state.editorState, event.shiftKey ? -1 : 1, 4);
+      const editorState = changeDepth(
+        this.state.editorState,
+        event.shiftKey ? -1 : 1,
+        4
+      );
       if (editorState && editorState !== this.state.editorState) {
         this.onChange(editorState);
         event.preventDefault();
@@ -214,13 +233,17 @@ export default class WysiwygEditor extends Component {
 
   onChange: Function = (editorState: Object): void => {
     const { readOnly, onEditorStateChange } = this.props;
-    if (!readOnly &&
-      !(getSelectedBlocksType(editorState) === 'atomic' &&
-      editorState.getSelection().isCollapsed)) {
+    if (
+      !readOnly &&
+      !(
+        getSelectedBlocksType(editorState) === "atomic" &&
+        editorState.getSelection().isCollapsed
+      )
+    ) {
       if (onEditorStateChange) {
         onEditorStateChange(editorState, this.props.wrapperId);
       }
-      if (!hasProperty(this.props, 'editorState')) {
+      if (!hasProperty(this.props, "editorState")) {
         this.setState({ editorState }, this.afterChange(editorState));
       } else {
         this.afterChange(editorState);
@@ -233,28 +256,36 @@ export default class WysiwygEditor extends Component {
   };
 
   setEditorReference: Function = (ref: Object): void => {
+    if (this.props.editorRef) {
+      this.props.editorRef(ref);
+    }
     this.editor = ref;
   };
 
   getCompositeDecorator = (): void => {
-    const decorators = [...this.props.customDecorators, getLinkDecorator({
-      showOpenOptionOnHover: this.state.toolbar.link.showOpenOptionOnHover,
-    })];
+    const decorators = [
+      ...this.props.customDecorators,
+      getLinkDecorator({
+        showOpenOptionOnHover: this.state.toolbar.link.showOpenOptionOnHover
+      })
+    ];
     if (this.props.mention) {
-      decorators.push(...getMentionDecorators({
-        ...this.props.mention,
-        onChange: this.onChange,
-        getEditorState: this.getEditorState,
-        getSuggestions: this.getSuggestions,
-        getWrapperRef: this.getWrapperRef,
-        modalHandler: this.modalHandler,
-      }));
+      decorators.push(
+        ...getMentionDecorators({
+          ...this.props.mention,
+          onChange: this.onChange,
+          getEditorState: this.getEditorState,
+          getSuggestions: this.getSuggestions,
+          getWrapperRef: this.getWrapperRef,
+          modalHandler: this.modalHandler
+        })
+      );
     }
     if (this.props.hashtag) {
       decorators.push(getHashtagDecorator(this.props.hashtag));
     }
     return new CompositeDecorator(decorators);
-  }
+  };
 
   getWrapperRef = () => this.wrapper;
 
@@ -278,31 +309,41 @@ export default class WysiwygEditor extends Component {
 
   isImageAlignmentEnabled = () => this.state.toolbar.image.alignmentEnabled;
 
-  createEditorState = (compositeDecorator) => {
+  createEditorState = compositeDecorator => {
     let editorState;
-    if (hasProperty(this.props, 'editorState')) {
+    if (hasProperty(this.props, "editorState")) {
       if (this.props.editorState) {
-        editorState = EditorState.set(this.props.editorState, { decorator: compositeDecorator });
+        editorState = EditorState.set(this.props.editorState, {
+          decorator: compositeDecorator
+        });
       }
-    } else if (hasProperty(this.props, 'defaultEditorState')) {
+    } else if (hasProperty(this.props, "defaultEditorState")) {
       if (this.props.defaultEditorState) {
-        editorState = EditorState.set(
-          this.props.defaultEditorState,
-          { decorator: compositeDecorator },
-        );
+        editorState = EditorState.set(this.props.defaultEditorState, {
+          decorator: compositeDecorator
+        });
       }
-    } else if (hasProperty(this.props, 'contentState')) {
+    } else if (hasProperty(this.props, "contentState")) {
       if (this.props.contentState) {
         const contentState = convertFromRaw(this.props.contentState);
-        editorState = EditorState.createWithContent(contentState, compositeDecorator);
+        editorState = EditorState.createWithContent(
+          contentState,
+          compositeDecorator
+        );
         editorState = EditorState.moveSelectionToEnd(editorState);
       }
-    } else if (hasProperty(this.props, 'defaultContentState')
-      || hasProperty(this.props, 'initialContentState')) {
-      let contentState = this.props.defaultContentState || this.props.initialContentState;
+    } else if (
+      hasProperty(this.props, "defaultContentState") ||
+      hasProperty(this.props, "initialContentState")
+    ) {
+      let contentState =
+        this.props.defaultContentState || this.props.initialContentState;
       if (contentState) {
         contentState = convertFromRaw(contentState);
-        editorState = EditorState.createWithContent(contentState, compositeDecorator);
+        editorState = EditorState.createWithContent(
+          contentState,
+          compositeDecorator
+        );
         editorState = EditorState.moveSelectionToEnd(editorState);
       }
     }
@@ -310,21 +351,50 @@ export default class WysiwygEditor extends Component {
       editorState = EditorState.createEmpty(compositeDecorator);
     }
     return editorState;
-  }
+  };
 
-  filterEditorProps = props => filter(props, [
-    'onChange', 'onEditorStateChange', 'onContentStateChange', 'initialContentState',
-    'defaultContentState', 'contentState', 'editorState', 'defaultEditorState', 'locale',
-    'localization', 'toolbarOnFocus', 'toolbar', 'toolbarCustomButtons', 'toolbarClassName',
-    'editorClassName', 'toolbarHidden', 'wrapperClassName', 'toolbarStyle', 'editorStyle',
-    'wrapperStyle', 'uploadCallback', 'onFocus', 'onBlur', 'onTab', 'mention', 'hashtag',
-    'ariaLabel', 'customBlockRenderFunc', 'customDecorators', 'handlePastedText'
-  ]);
+  filterEditorProps = props =>
+    filter(props, [
+      "onChange",
+      "onEditorStateChange",
+      "onContentStateChange",
+      "initialContentState",
+      "defaultContentState",
+      "contentState",
+      "editorState",
+      "defaultEditorState",
+      "locale",
+      "localization",
+      "toolbarOnFocus",
+      "toolbar",
+      "toolbarCustomButtons",
+      "toolbarClassName",
+      "editorClassName",
+      "toolbarHidden",
+      "wrapperClassName",
+      "toolbarStyle",
+      "editorStyle",
+      "wrapperStyle",
+      "uploadCallback",
+      "onFocus",
+      "onBlur",
+      "onTab",
+      "mention",
+      "hashtag",
+      "ariaLabel",
+      "customBlockRenderFunc",
+      "customDecorators",
+      "handlePastedText"
+    ]);
 
-  changeEditorState = (contentState) => {
+  changeEditorState = contentState => {
     const newContentState = convertFromRaw(contentState);
     let { editorState } = this.state;
-    editorState = EditorState.push(editorState, newContentState, 'insert-characters');
+    editorState = EditorState.push(
+      editorState,
+      newContentState,
+      "insert-characters"
+    );
     editorState = EditorState.moveSelectionToEnd(editorState);
     return editorState;
   };
@@ -337,7 +407,7 @@ export default class WysiwygEditor extends Component {
 
   handleKeyCommand: Function = (command: Object): boolean => {
     const { editorState, toolbar: { inline } } = this.state;
-    if(inline && inline.options.indexOf(command) >= 0) {
+    if (inline && inline.options.indexOf(command) >= 0) {
       const newState = RichUtils.handleKeyCommand(editorState, command);
       if (newState) {
         this.onChange(newState);
@@ -360,16 +430,24 @@ export default class WysiwygEditor extends Component {
   };
 
   handlePastedText = (text, html) => {
-    if (this.props.handlePastedText &&
-      this.props.handlePastedText(text, html, editorState, this.onChange)) {
-        return true;
+    const { editorState } = this.state;
+
+    if (this.props.handlePastedText) {
+      return this.props.handlePastedText(
+        text,
+        html,
+        editorState,
+        this.onChange
+      );
     }
-    const { editorState } = this.state;    
-    return handlePastedText(text, html, editorState, this.onChange);
-  }
+    if (!this.props.stripPastedStyles) {
+      return handlePastedText(text, html, editorState, this.onChange);
+    }
+    return false;
+  };
 
   preventDefault: Function = (event: Object) => {
-    if (event.target.tagName === 'INPUT') {
+    if (event.target.tagName === "INPUT" || event.target.tagName === "LABEL") {
       this.focusHandler.onInputMouseDown();
     } else {
       event.preventDefault();
@@ -377,11 +455,7 @@ export default class WysiwygEditor extends Component {
   };
 
   render() {
-    const {
-      editorState,
-      editorFocused,
-      toolbar,
-    } = this.state;
+    const { editorState, editorFocused, toolbar } = this.state;
     const {
       locale,
       localization: { locale: newLocale, translations },
@@ -395,48 +469,58 @@ export default class WysiwygEditor extends Component {
       editorStyle,
       wrapperStyle,
       uploadCallback,
-      ariaLabel,
+      ariaLabel
     } = this.props;
 
     const controlProps = {
       modalHandler: this.modalHandler,
       editorState,
       onChange: this.onChange,
-      translations: { ...localeTranslations[locale || newLocale], ...translations },
+      translations: {
+        ...localeTranslations[locale || newLocale],
+        ...translations
+      }
     };
-    const toolbarShow = !toolbarHidden &&
-      (editorFocused || this.focusHandler.isInputFocused() || !toolbarOnFocus);
+    const toolbarShow =
+      editorFocused || this.focusHandler.isInputFocused() || !toolbarOnFocus;
     return (
       <div
         id={this.wrapperId}
-        className={classNames(wrapperClassName, 'rdw-editor-wrapper')}
+        className={classNames(wrapperClassName, "rdw-editor-wrapper")}
         style={wrapperStyle}
         onClick={this.modalHandler.onEditorClick}
         onBlur={this.onWrapperBlur}
         aria-label="rdw-wrapper"
       >
-        <div
-          className={classNames('rdw-editor-toolbar', toolbarClassName)}
-          style={{ visibility: toolbarShow ? 'visible' : 'hidden', ...toolbarStyle }}
-          onMouseDown={this.preventDefault}
-          aria-label="rdw-toolbar"
-          aria-hidden={(!editorFocused && toolbarOnFocus).toString()}
-          onFocus={this.onToolbarFocus}
-        >
-          {toolbar.options.map((opt, index) => {
-            const Control = Controls[opt];
-            const config = toolbar[opt];
-            if (opt === 'image' && uploadCallback) {
-              config.uploadCallback = uploadCallback;
-            }
-            return <Control key={index} {...controlProps} config={config} />;
-          })}
-          {toolbarCustomButtons && toolbarCustomButtons.map((button, index) =>
-            React.cloneElement(button, { key: index, ...controlProps }))}
-        </div>
+        {!toolbarHidden && (
+          <div
+            className={classNames("rdw-editor-toolbar", toolbarClassName)}
+            style={{
+              visibility: toolbarShow ? "visible" : "hidden",
+              ...toolbarStyle
+            }}
+            onMouseDown={this.preventDefault}
+            aria-label="rdw-toolbar"
+            aria-hidden={(!editorFocused && toolbarOnFocus).toString()}
+            onFocus={this.onToolbarFocus}
+          >
+            {toolbar.options.map((opt, index) => {
+              const Control = Controls[opt];
+              const config = toolbar[opt];
+              if (opt === "image" && uploadCallback) {
+                config.uploadCallback = uploadCallback;
+              }
+              return <Control key={index} {...controlProps} config={config} />;
+            })}
+            {toolbarCustomButtons &&
+              toolbarCustomButtons.map((button, index) =>
+                React.cloneElement(button, { key: index, ...controlProps })
+              )}
+          </div>
+        )}
         <div
           ref={this.setWrapperReference}
-          className={classNames(editorClassName, 'rdw-editor-main')}
+          className={classNames(editorClassName, "rdw-editor-main")}
           style={editorStyle}
           onClick={this.focusEditor}
           onFocus={this.onEditorFocus}
@@ -457,7 +541,7 @@ export default class WysiwygEditor extends Component {
             handlePastedText={this.handlePastedText}
             blockRendererFn={this.blockRendererFn}
             handleKeyCommand={this.handleKeyCommand}
-            ariaLabel={ariaLabel || 'rdw-editor'}
+            ariaLabel={ariaLabel || "rdw-editor"}
             blockRenderMap={blockRenderMap}
             {...this.editorProps}
           />