+++ stories/Basic/index.js
@@ -5,6 +5,7 @@ import { Editor } from '../../src';
 
 const Basic = () => (<div className="rdw-storybook-root">
   <Editor
+    ref={(ref) => {console.log('hey ref', ref.focusEditor())}}
     toolbarClassName="rdw-storybook-toolbar"
     wrapperClassName="rdw-storybook-wrapper"
     editorClassName="rdw-storybook-editor"